Update node-install
arm64 binaries are not working on the aarch64 test version of the Raspbian Buster arm_64bit=1 in config.txt, and after running sudo rpi-update.  Reverting back to forcing the Raspberry pi 4 to install, running in 64bit kernel mode, the armv7l for now, so that this will not break people's installation/updating node on the pi 4 specifically.  This might break other computers that report aarch64, but since this script is for Raspberry Pi's specifically, I'm going to support the Pi4 for now. Maybe the a fully supported release of Raspbian 64 will change this back to allowing the arm64 binaries to install.